pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krakey"
version = "0.1.0"
description = "KrakeyBot — autonomous cognitive agent that maintains presence via continuous heartbeat."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Samuel" }]
keywords = ["agent", "llm", "autonomous", "cognitive", "heartbeat", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "aiosqlite>=0.19",
    "sqlite-vec>=0.1.6",
    "igraph>=0.11",
    "leidenalg>=0.10",
    "ddgs>=9.0",
    "pyautogui>=0.9",
    "fastapi>=0.115",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30",
    "tiktoken>=0.7",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
krakey = "krakey.cli:main"

[project.urls]
Homepage = "https://github.com/Arrosam/KrakeyBot"

[tool.setuptools.packages.find]
where = ["."]
include = ["krakey*"]

[tool.setuptools.package-data]
"krakey.memory" = ["schemas.sql"]
"krakey.prompt" = ["dna.txt"]
"krakey.plugins.dashboard" = ["meta.yaml", "static/**/*"]
"krakey.plugins.duckduckgo_search" = ["meta.yaml"]
"krakey.plugins.hypothalamus" = ["meta.yaml"]
"krakey.plugins.in_mind_note" = ["meta.yaml"]
"krakey.plugins.recall" = ["meta.yaml"]
"krakey.plugins.telegram" = ["meta.yaml"]
